requirements.txt
google-api-python-client==2.125.0
google-auth==2.29.0
openai==1.30.1
requests==2.31.0
